requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pydantic==2.3.0
pydantic-settings==2.0.3
httpx==0.25.0
python-dotenv==1.0.0
opencv-python==4.8.1.78
pillow==10.1.0
aiofiles==23.2.1
